Move entity specific logic from instantiateRuntimeValue to nativeInstantiate
